Refactor server side telemetry
