+++ msal/src/main/java/com/microsoft/identity/client/SingleAccountPublicClientApplication.java
@@ -32,7 +32,6 @@ import androidx.annotation.WorkerThread;
 import com.microsoft.identity.client.exception.MsalClientException;
 import com.microsoft.identity.client.exception.MsalException;
 import com.microsoft.identity.client.internal.AsyncResult;
-import com.microsoft.identity.client.internal.PublicApiId;
 import com.microsoft.identity.client.internal.controllers.BrokerMsalController;
 import com.microsoft.identity.client.internal.controllers.MSALControllerFactory;
 import com.microsoft.identity.client.internal.controllers.MsalExceptionAdapter;
@@ -51,7 +50,7 @@ import com.microsoft.identity.common.internal.request.OperationParameters;
 import com.microsoft.identity.common.internal.result.ILocalAuthenticationResult;
 import com.microsoft.identity.common.internal.result.MsalBrokerResultAdapter;
 import com.microsoft.identity.common.internal.result.ResultFuture;
-import com.microsoft.identity.common.internal.servertelemetry.ServerTelemetry;
+import com.microsoft.identity.common.internal.servertelemetry.PublicApiId;
 
 import java.util.List;
 
@@ -95,8 +94,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
     public void getCurrentAccountAsync(@NonNull final CurrentAccountCallback callback) {
         final String methodName = ":getCurrentAccount";
         final PublicClientApplicationConfiguration configuration = getConfiguration();
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_GET_CURRENT_ACCOUNT_ASYNC_WITH_CALLBACK);
-
 
         try {
             if (mIsSharedDevice) {
@@ -141,6 +138,7 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
 
             );
 
+            command.setPublicApiId(PublicApiId.GET_CURRENT_ACCOUNT_ASYNC);
             ApiDispatcher.getAccounts(command);
 
         } catch (MsalClientException clientException) {
@@ -153,8 +151,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
         throwOnMainThread("getCurrentAccount");
 
         final ResultFuture<AsyncResult<CurrentAccountResult>> future = new ResultFuture<>();
-
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_GET_CURRENT_ACCOUNT);
 
         getCurrentAccountAsync(new CurrentAccountCallback() {
             @Override
@@ -222,7 +218,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
                        @NonNull final String loginHint,
                        @NonNull final String[] scopes,
                        @NonNull final AuthenticationCallback callback) {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_SIGN_IN);
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount != null) {
             callback.onError(new MsalClientException(MsalClientException.INVALID_PARAMETER));
@@ -297,7 +292,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
 
     @Override
     public void signOut(@NonNull final SignOutCallback callback) {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_SIGN_OUT_WITH_CALLBACK);
         final PublicClientApplicationConfiguration configuration = getConfiguration();
 
         final MultiTenantAccount persistedCurrentAccount = getPersistedCurrentAccount();
@@ -339,6 +333,7 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
                     }
             );
 
+            command.setPublicApiId(PublicApiId.SIGN_OUT);
             ApiDispatcher.removeAccount(command);
         } catch (final MsalClientException clientException) {
             callback.onError(clientException);
@@ -351,8 +346,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
         throwOnMainThread("signOut");
 
         final ResultFuture<AsyncResult<Boolean>> future = new ResultFuture<>();
-
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_SIGN_OUT);
 
         signOut(new SignOutCallback() {
             @Override
@@ -459,7 +452,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
     public void acquireToken(@NonNull final Activity activity,
                              @NonNull final String[] scopes,
                              @NonNull final AuthenticationCallback callback) {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_ACQUIRE_TOKEN_WITH_ACTIVITY_SCOPES_CALLBACK);
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount == null) {
             callback.onError(new MsalClientException(MsalClientException.NO_CURRENT_ACCOUNT));
@@ -482,7 +474,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
 
     @Override
     public void acquireToken(@NonNull final AcquireTokenParameters acquireTokenParameters) {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_ACQUIRE_TOKEN_WITH_PARAMETERS);
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount != null) {
             // If the account exists, overwrite Account and ignore loginHint.
@@ -497,8 +488,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
     public void acquireTokenSilentAsync(@NonNull final String[] scopes,
                                         @NonNull final String authority,
                                         @NonNull final SilentAuthenticationCallback callback) {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_ACQUIRE_TOKEN_SILENT_ASYNC_WITH_SCOPES_AUTHORITY_CALLBACK);
-
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount == null) {
             callback.onError(new MsalClientException(MsalClientException.NO_CURRENT_ACCOUNT));
@@ -518,7 +507,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
     @WorkerThread
     public IAuthenticationResult acquireTokenSilent(@NonNull final String[] scopes,
                                                     @NonNull final String authority) throws MsalException, InterruptedException {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_ACQUIRE_TOKEN_SILENT_WITH_SCOPES_AUTHORITY);
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount == null) {
             MsalClientException exception = new MsalClientException(MsalClientException.NO_CURRENT_ACCOUNT);
@@ -530,7 +518,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
 
     @Override
     public void acquireTokenSilentAsync(@NonNull final AcquireTokenSilentParameters acquireTokenSilentParameters) {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_ACQUIRE_TOKEN_SILENT_ASYNC_WITH_PARAMETERS);
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount == null) {
             acquireTokenSilentParameters.getCallback().onError(new MsalClientException(MsalClientException.NO_CURRENT_ACCOUNT));
@@ -545,7 +532,6 @@ public class SingleAccountPublicClientApplication extends PublicClientApplicatio
 
     @Override
     public IAuthenticationResult acquireTokenSilent(@NonNull final AcquireTokenSilentParameters acquireTokenSilentParameters) throws InterruptedException, MsalException {
-        ServerTelemetry.emitApiId(PublicApiId.SINGLE_ACCOUNT_PCA_ACQUIRE_TOKEN_SILENT_WITH_PARAMETERS);
         final IAccount persistedAccount = getPersistedCurrentAccount();
         if (persistedAccount == null) {
             throw new MsalClientException(MsalClientException.NO_CURRENT_ACCOUNT);
